tests: fix sonarcloud duplication error by combining test triggers
